+++ samples/My.Hr/My.Hr.Business/Data/GenderConfiguration.cs
@@ -15,5 +15,10 @@ public class GenderConfiguration : IEntityTypeConfiguration<Gender>
         entity.Ignore(p => p.EndDate);
         entity.Ignore(p => p.StartDate);
         entity.Ignore(p => p.Description);
+        entity.Ignore(p => p.IsReadOnly);
+        entity.Ignore(p => p.IsValid);
+        entity.Ignore(p => p.IsChanged);
+        entity.Ignore(p => p.IsInitial);
+        entity.Ignore(p => p.NotifyChangesWhenSameValue);
     }
 }