+++ samples/My.Hr/My.Hr.Business/Models/Employee.cs
@@ -28,7 +28,7 @@ public class Employee : IIdentifier<Guid>, IETag
     /// <summary>
     /// Gets or sets the 'GenderCode' column value.
     /// </summary>
-    public string? Gender { get; set; }
+    public Gender? Gender { get; set; }
 
     /// <summary>
     /// Gets or sets the 'Birthday' column value.