+++ samples/My.Hr/My.Hr.Business/Models/Gender.cs
@@ -1,8 +1,10 @@
 ﻿using CoreEx.RefData;
-using CoreEx.RefData.Models;
 
 namespace My.Hr.Business.Models;
 
-public class Gender : ReferenceDataBase<Guid> { }
+public class Gender : ReferenceDataBase<Guid, Gender> 
+{
+    public static implicit operator Gender?(string? code) => ConvertFromCode(code);
+}
 
 public class GenderCollection : ReferenceDataCollection<Guid, Gender, GenderCollection> { }