+++ samples/My.Hr/My.Hr.Business/Models/UsState.cs
@@ -1,5 +1,8 @@
 namespace My.Hr.Business.Models;
 
-public class USState : ReferenceDataBase<Guid> { }
+public class USState : ReferenceDataBase<Guid, USState>
+{
+    public static implicit operator USState?(string? code) => ConvertFromCode(code);
+}
 
 public class USStateCollection : ReferenceDataCollection<Guid, USState, USStateCollection> { }